updated child medical history
