+++ flourish_child/admin/child_medical_history_admin.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 from edc_fieldsets.fieldlist import Insert
 from edc_fieldsets.fieldsets_modeladmin_mixin import FormLabel
 from edc_model_admin import audit_fieldset_tuple
+from edc_fieldsets.fieldsets import Fieldsets
 
 
 from ..admin_site import flourish_child_admin
@@ -39,9 +40,9 @@ class ChildMedicalHistoryAdmin(ChildCrfModelAdminMixin, admin.ModelAdmin):
         "chronic_since": admin.VERTICAL,
         "med_history_changed": admin.VERTICAL,
         "current_hiv_status": admin.VERTICAL,
-        "is_pregnant": admin.VERTICAL,
-        "is_lmp_date_estimated": admin.VERTICAL,
+        "preg_test_performed": admin.VERTICAL,
         "pregnancy_test_result": admin.VERTICAL,
+        "is_lmp_date_estimated": admin.VERTICAL,
     }
 
     filter_horizontal = ("child_chronic",)
@@ -78,13 +79,49 @@ class ChildMedicalHistoryAdmin(ChildCrfModelAdminMixin, admin.ModelAdmin):
         ),
         "child_pool_schedule1": Insert("med_history_changed", after="report_datetime"),
         "female_above_12": Insert(
-            "is_pregnant",
+            "preg_test_performed",
+            "pregnancy_test_result",
             "last_menstrual_period",
             "is_lmp_date_estimated",
-            "pregnancy_test_result",
             after="current_hiv_status",
         ),
     }
+    
+    def get_fieldsets_update(self, request, obj=None):
+        """Returns fieldsets after modifications declared in
+        "conditional" dictionaries.
+        """
+        fieldsets = super().get_fieldsets(request, obj=obj)
+        fieldsets = Fieldsets(fieldsets=fieldsets)
+        keys = self.get_keys(request, obj)
+        for key in keys:
+            fieldset = self.conditional_fieldsets.get(key)
+            if fieldset:
+                try:
+                    fieldset = tuple(fieldset)
+                except TypeError:
+                    fieldset = (fieldset,)
+                for f in fieldset:
+                    fieldsets.add_fieldset(fieldset=f)
+            fieldlist = self.conditional_fieldlists.get(key)
+            if fieldlist:
+                try:
+                    fieldsets.insert_fields(
+                        *fieldlist.insert_fields,
+                        insert_after=fieldlist.insert_after,
+                        section=fieldlist.section)
+                except AttributeError:
+                    pass
+                try:
+                    fieldsets.remove_fields(
+                        *fieldlist.remove_fields,
+                        section=fieldlist.section)
+                except AttributeError:
+                    pass
+        fieldsets = self.update_fieldset_for_form(
+            fieldsets, request)
+        fieldsets.move_to_end(self.fieldsets_move_to_end)
+        return fieldsets.fieldsets
 
     def get_fieldsets(self, request, obj=None):
         return self.get_fieldsets_update(request, obj)
